Fix scaling-dependence of termination criterion
The termination criterion for bisecting alpha implicitly assumes the
optimal alpha to be in the order of magnitude of 1.0. For larger values
of alpha, lower and upper can never get close enough, meaning the
bisection never terminates successfully. For smaller values of alpha,
the bisection may terminate much too early.

This fixes that by checking the relative instead of the absolute
difference using numpy.isclose. Also, this adds tests with a large and
a tiny optimal alpha that would have failed with the previous
termination criterion (they also wouldve failed for much more moderate
scaling factors like 100).